backend/src/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
Django==3.0.6
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.2.2
pytz==2020.1
sqlparse==0.3.1
